requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
bitarray==2.7.3
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
cytoolz==0.12.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
Flask==2.2.3
fonttools==4.39.3
frozenlist==1.3.3
gunicorn==20.1.0
hexbytes==0.3.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lru-dict==1.1.8
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==2.0.0
parsimonious==0.9.0
Pillow==9.5.0
prettytable==3.6.0
protobuf==4.22.1
pycryptodome==3.17
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rlp==3.0.0
seaborn==0.12.2
six==1.16.0
soupsieve==2.4
toolz==0.12.0
tweepy==4.13.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
web3==6.0.0
websockets==11.0
Werkzeug==2.2.3
yarl==1.8.2